Get rid of some ncurses memory leaks. This is really ncurses's fault, not ours, but since most of the leaks can easily be removed by calling newterm and delscreen, it seemed reasonable to do so.
I've discussed the rest of the leaks (detected by valgrind) with the
maintainer, and he doesn't want to or can't fix the remainder: one
leak would "require changing the ABI", and another is a
setbuffer/setvbuf buffer that "tends to cause core dumps" if it's
freed.

I'm inclined to add a suppressions file to valgrind for these that
ignores them.