Get rid of castr and replace it with const astr, which is self-documenting, and used rarely enough that having a specific alias is not worthwhile.
Get rid of ENLARGE_FACTOR which wasn't used.

Add static astr_pos(), which turns negative positions into positive
for those routines that take a pos parameter. Instead of truncating
the given pos to the actual size of the string, it asserts that it is
in range; I'll see if this causes problems and take appropriate
action. It does seem preferable to allowing bad positions, though.

Add astr_char, to return the address of a given character.

Remove astr_last_char (replaced with *astr_char(as ,-1)).

Moved more documentation to astr.h from astr.3.