DON'T PANIC! I have tagged the tree with the tag "fontlock-apocalypse" before committing this set of changes.
Remove font-lock modes (N.B. *not* font-lock support!). The font-lock
system in zile was a Bad Thing:

 * It was a maintenance burden because of conditional compilation and
   configure options.
 * It was a lot of complicated code. Several recent bugs have been
   font-lock bugs.
 * In terms of functionality (and to a lesser extent, code size) it
   was a bloat on an editor that should be small, clean and simple.
 * It added a dependency on gperf (admittedly not a big problem).

I have retained font-lock support, and made it always-on. Regex-based
font-lock should be implemented, similar to that in many other
editors. Preferably, we should adopt an existing regex patterns format
(e.g. that of NEdit) so that we have access to a wide range of
fontlock modes, and users don't have to make yet another crufty set.

Alternatively, it may be better in an editor the size of zile to
remove font-lock altogether.

Remove experimental Lua support. For similar reasons to those above,
it was a bad idea.