More cleanup:
1. Removed [EOB] marker. This is not an Emacs feature, and most
editors (rightly, in my view) don't have it. It's easy to know when
you're at the end of the buffer by a combination of the status line
and the down key!

2. Removed some more vestiges of font-lock: in particular, the
font-lock variables and the OUTCH macro.

3. Start isolating the ncurses code. To do this, add stub functions
and variables to term_ncurses.c for ncurses functions. The types and
macros are going to be a bit more annoying, and some code will
obviously have to stay in this directory. ncurses_minibuf.c is now
"ncurses free", i.e. doesn't need to #include ncurses.h.