Remove some duplicate functions.
The situation is now looking a little messy: there's an API in
term_ncurses.h, and there's another one offered by the Terminal type.
I intend to get rid of the Terminal type: it's cute, but it ain't
useful, as we will end up with the terminal type being a compile-time,
not a run-time switch. Preferably with only one setting :-)