Get rid of term_minibuf_clear, and replace it with term_minibuf_write(""), which is functionally identical.
Get rid of term_getyx, as it can't be implemented in termcap (I was
using the ncurses getyx call, which returned nonsense results).

Fix a bug in term_addnstr in term_termcap.c.


Comment:

It seems that not calling term_getyx to save the position of the
cursor when writing to the minibuffer has caused no ill effects, but I
could be wrong; something to watch out for...

The display now seems basically to work under termcap, as the changes
checked in here fix the minibuffer.