Make the tests be built by make check.
Make two non-API functions, one in astr.c and one in alist.c, static.

Fix some minor bugs in htable.c revealed by the tests, and remove the
unused API function htable_remove. If this is needed in future it
should be implemented better (e.g. by having htable_store remove a
bucket if the value is set to NULL).

Add the test programs to .cvsignore.