Add a lisp interpreter, code taken from lithp 0.6 (http://www.cis.rit.edu/~jerry/Software/lithp/).
Since the original was written with noweb, I've tangled the code,
stripped out the #line directives, and reformatted it.

I've also done something a bit naughty: I've fixed up most of the gcc
warnings, and done a little tidying to make that work, hence the code
is not quite the same as lithp 0.6. I'm checking it in now before I
really start changing stuff.

At the moment the code is simply compiled into Zile, but not called
anywhere.

Justification: having a Lisp interpreter will simplify and unify a lot
of things. First, it will handle the .zilerc file (and hence get rid
of rc.c and a lot of variables.c). Secondly, it will simplify adding
key rebinding. Thirdly, it's entirely in line with Emacs emulation. It
will similarly simplify the handling of macros, and allow Zile to be
extended at last by the user. Fourthly, it will help implement key
maps, which are needed to remove horrible code duplication in the
minibuffer code. Finally, some of the current C code will be rewritten
in Lisp, and it will hence improve stability (as Lisp code won't
crash, right?!).

I will try very hard to reduce the binary size to at most the same as
it was before I added the interpreter, which is only 2,000 lines of
code. It adds about 50Kb to the unstrippped binary; I have already
removed nearly 20Kb since the current 2.0 beta, so that leaves another
30Kb to remove. I'll try to do the same in the source code, but that's
less important (and less likely). Anyway, Zile should get no more
complex than it is at present until some other major feature is added
(the only one planned is Unicode support).

I have yet to put in proper copyright notices in the files from lithp
and proper acknowledgement for this major contribution in the docs.
The lithp code will also doubtless be heavily modified (and reduced!).