Remove garbage_collect, which was just a stub, and the corresponding eval_cb_nothing.
Tidy up eval.c, formatting it like the rest of the Zile code, and
tidying up a lot of the code.

Make leWipe not free leNIL, and add leReallyWipe, which does free
leNIL. Could use this for other constants, but really need a gc.

Remove tests of functions that no longer exist from test3.el.