Add a buffer-local variable list to the Buffer struct, and move the two buffer-local variables (fill-column and tab-width) into it. Remove tab_width and fill_column.
Add tab_width to return a "safe" tab width, i.e. at least 1.

Add get_variable_bp to get a variable for a particular buffer,
get_variable_number to get a variable's value as a number, and
get_variable_number_bp.

Add variableSetNumber.

Reduce the number of calls to read the value of tab-width where this
is easy.

Make describe-variable get the value for the current buffer for a
buffer-local variable.

Improve doc strings for newline-and-indent and indent-command.