Move mkdoc from doc to src directory and rename mkfuncs. Simplify the code to use only standard functions (no other Zile code). Instead of reading tbl_funcs.h, make it generate it. This means that quotes in docstrings must be escaped; there was only one instance of this.
Don't make AUTODOC, but compile docstrings into the zile binary.

Move tbl_funcs.h to tbl_bind.h, and simplify to include only bound
commands.

Make zlint able to cope with unsorted tbl_funcs.h.

Remove tutorial (which is not really useful for the intended target
audience), and HELP file (which is not the same as that shown in Emacs
now).

Remove mention of dotsample.zile from zile(1), and thereby remove need
for zile.1.in.