Fix shell-command and shell-command-on-region for various cases, and factor out much of the code into pipe_command.
Update shell-command's docstring from Emacs's.

To allow output of shell-command[-on-region] to be available in
*Shell-Output* buffer even when the buffer is not shown (because the
output is short enough to go in the echo area), add an argument to
write_temp_buffer to specify whether the temp buffer should actually
be shown or not.

Add test cases for shell-command and shell-command-on-region.