Remove exchange_point_and_mark in favour of the Zile command.
Fix up transpose_subr. In particular, go back to calling the primitive
functions, which means we have to pass both backward and forward
flavours, because we don't want to raise an error when they fail. In
order to make this work, move the transpose functions after all the
static functions they now call. Not fully tested yet, but working a
lot better than it was before!