Rename "tests" directory to "lisp-tests" for clarity.
Add execute-kbd-macro to allow tests requiring keyboard input to be
written. Get the rest of the tests working.

As a side-effect, change the implementation of macros to use
gl_list_t's rather than dynamically-allocated size_t arrays with
counts to hold the key list, as this works better with the functions
used for execute-kbd-macro (and is cleaner in any case).