In order to generate the string "ZILE" for emergency saves, generalise the static function recase called by line_replace_text to astr_replace.
Tidy up setcase_region slightly.

Remove the newlen argument from line_replace_text, as although this in
practice results in extra calls to strlen, it's good to signal that
the newstr argument is always a NULL-terminated string, and is one
more step on the way to using astr's rather than char *'s.