Replace astr_char with astr_get (which returns the character rather than a pointer to it).
Use astr_cstr for remaining uses of astr_char.

Remove out-of-date explanation of string positions from astr.h.

Make second arg of make_char_printable an int, as it should be.

In line_replace_text, avoid a bug caused by redeclaring case_type, so
that if it was updated the value was not available at the end of the
function.

Rewrite calculate_start_column to use string offsets rather than
pointers.