Start making tests Emacs-compatible.
run-lisp-tests: fix assignment of args so it is always made; copy
input file for both tests, not just the first.

Add macros for arguments that can be either non-interactively supplied
in Lisp or interactively supplied by universal argument, and start
using it with those commands that should take an argument
non-interactively.

Start making tests work on Emacs.