Reimplement buffers as a single string, into which Lines are pointers. Lines are treated now as non-circular lists, implemented by methods rather than struct members.
The code currently has holes: in particular, adjust_markers can crash
when the current line number of a marker is wrong because of deletion.
Rather than fix this it seems more sensible to proceed to reimplement
Markers as simple offsets into a buffer, but also a good idea to check
in what I’ve done so far!