[configury] move to non-recursive make and canonicalize directory layout
Instead of putting *everything* in `src/', move documentation into `doc/',
auxilliary build scripts into `build-aux/' and tests into `tests/' in
keeping with the layout of most GNU packages.  Updated README to explain
where everything is, as is the tradition.

Then updated the configury to perform a non-recursive build like the lua
branch, with the exception of `lib/' which contains a gnulib-tool
generated Makefile.am that is not compatible with a non-recursive scheme.