Improve efficiency of replace_estr by not moving or writing more data than absolutely necessary: use new estr_replace_estr and estr_len, and update cur_bp->pt directly.
Replace estr_replace with estr_replace_estr, which, like
astr_replace_nstr, overwrites rather than inserting.

Add estr_lines, which counts the number of lines in an estr, and
estr_len, which returns the length of an estr assuming conversion to a
given line ending.

Make set_buffer_pt static, as it is not used outside buffer.c.