tests: fix tests/interactive/backward-sexp.el for Emacs.
Emacs behaves strangely in batch mode, adding a paren on the end
of the second line of a file for (execute-kbd-macro "\M-<\C-e)")
so put the sexp closing paren at the end of the last line
instead, which even batch mode does correctly.
* tests/interactive/backward-sexp.el: Put the closing paren
before the final newline of the test file.
* tests/interactive/backward-sexp.output: Adjust.